Fix signedness of types in python getitem/setitem, which allow negative indices. Also avoid serializing size_t (I think cereal actually does the right thing here, but it's good to be explicit).

+++ coordinateutils/include/coordinateutils/FlatSkyMap.h
@@ -122,7 +122,7 @@ private:
 
 	DenseMapData *dense_;
 	SparseMapData *sparse_;
-	size_t xpix_, ypix_;
+	uint64_t xpix_, ypix_;
 
 	SET_LOGGER("FlatSkyMap");
 };
